% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidData.R
\docType{data}
\name{banana}
\alias{banana}
\alias{banana-bc.test}
\alias{banana-bc.train}
\alias{banana-mc.test}
\alias{banana-mc.train}
\title{\code{banana-bc.train}, \code{banana-bc.test}
\code{banana-mc.train}, and \code{banana-mc.test}}
\description{
Generated data set having a binary or 4-level Y variable and
a two-dimensional X (first two levels resemble bananas).
Both the train and the test set have 2000 samples in the binary case,
and 4000 in the multi-class case.
They were generated by the authors and their collaborators.
}

