% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidSVM.R
\name{command-args}
\alias{command-args}
\title{liquidSVM command line options}
\description{
\strong{Should only be used by experts!}
liquidSVM command line tools \code{svm-train}, \code{svm-select}, and \code{svm-test}
can be used by more advanced users to get the most advanced use.
These three tools have command line arguments and those can be used from R as well.
}
\examples{
\dontrun{
reg <- liquidData('reg-1d')
model <- init.liquidSVM(Y~., reg$train)
trainSVMs(model, command.args=list(L=2, T=2, d=1))
selectSVMs(model, command.args=list(R=0,d=2))
result <- test(model, reg$test, command.args=list(T=1, d=0))
}
}

