% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidSVM.R
\name{compilationInfo}
\alias{compilationInfo}
\title{Compilation information: whether the library was compiled using SSE2 or even AVX.}
\usage{
compilationInfo()
}
\value{
character with the information.
}
\description{
Compilation information: whether the library was compiled using SSE2 or even AVX.
}

