% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{kern}
\alias{kern}
\title{Calculates the kernel matrix.}
\usage{
kern(data, gamma = 1, type = c("gaussian.rbf", "poisson"), threads = 1)
}
\arguments{
\item{data}{the data set}

\item{gamma}{the gamma-parameter}

\item{type}{kernel type to use: one of "gaussian.rbf","poisson"}

\item{threads}{how many threads to be used}
}
\value{
kernel matrix
}
\description{
Calculates the kernel matrix.
}
\examples{
kern(trees)
image(kern(trees, 2, "pois"))
}

