% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{liquidSVM-class}
\alias{liquidSVM-class}
\alias{liquidSVMclass}
\title{A Reference Class to represent a liquidSVM model.}
\description{
A Reference Class to represent a liquidSVM model.
}
\section{Fields}{

\describe{
\item{\code{cookie}}{this is used in C++ to access the model in memory}
}}

