% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.liquidSVM}
\alias{print.liquidSVM}
\title{Printing an SVM model.}
\usage{
\method{print}{liquidSVM}(x, ...)
}
\arguments{
\item{x}{the model to print}

\item{...}{other arguments to print.default}
}
\description{
Printing an SVM model.
}
\examples{
\dontrun{
s_iris <- svm(solver='hinge', Species ~ ., iris)  # multiclass classification
print(s_iris)
}
}
