% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidData.R
\docType{data}
\name{reg-1d}
\alias{reg-1d}
\alias{reg-1d.test}
\alias{reg-1d.train}
\title{\code{reg-1d.train} and \code{reg-1d.test}}
\description{
Generated data set having a continuous Y variable and
a one-dimensional X variable.
}
\details{
Both the train and the test set have 2000 samples.
They were generated by the authors and their collaborators.
}

