% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidSVM.R
\name{setDisplay}
\alias{setDisplay}
\title{Set display info mode that controls how much information is displayed
by liquidSVM C++ routines. Usually you will use \code{display=d} in \code{svm(...)} etc.}
\usage{
setDisplay(d = 1)
}
\arguments{
\item{d}{the display information}
}
\description{
Set display info mode that controls how much information is displayed
by liquidSVM C++ routines. Usually you will use \code{display=d} in \code{svm(...)} etc.
}

