% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidData.R
\name{write.liquidData}
\alias{write.liquidData}
\title{Write Smldata}
\usage{
write.liquidData(data, location = ".", label = 1, name = NULL,
  type = "csv")
}
\arguments{
\item{data}{the liquidData to write}

\item{location}{the location to write \code{name.train.csv} and \code{name.test.csv}}

\item{label}{the column with this index or this name will become the label column,
and be written as the first column.}

\item{name}{the name of the file. If \code{NULL} (default) then takes the \code{data$name}}

\item{type}{the format of output. At the moment only \code{"csv"} is supported.}
}
\description{
Write \code{liquidData} in such a way that it is understood by liquidSVM command line utilities.
}

